Sounds + Slopes w/ Blizzard
Sounds:
Looked at some of sounds we have and was messing around with playing sounds when you press a main menu button and using the song found in the google drive for main menu.

Slopes:
The Feature that I had been forgetting was Slopes! While I was going through the level scenes I noticed that scene 3 had a slope at the start and that the controller didn't handle it very well.
So added Slope Handling to the blizzard controller.

-TestLevel:
Added a slope under Blizzard at start of the level to show the slope handling working.