Sprites + Values
Sprites:
tweaked the pixels per unity on the Blizzard sprite because Blizzard looked slightly misshapen.

Values:
Tweaked some of the player movement values to be less forgiving (edge grabs and the power given by an edge grab).
Resized ground detection to line up with the size of the box collider.